Rename FlowNode -> FlowNodeInstance for consistency
