+++ src/apis/index.ts
@@ -2,7 +2,7 @@
 import * as correlationApi from './icorrelation_management_api';
 import * as emptyActivityApi from './iempty_activity_management_api';
 import * as eventApi from './ievent_management_api';
-import * as flowNodeApi from './iflow_node_management_api';
+import * as flowNodeInstanceApi from './iflow_node_instance_management_api';
 import * as kpiApi from './ikpi_management_api';
 import * as loggingApi from './ilogging_management_api';
 import * as manualTaskApi from './imanual_task_management_api';
@@ -15,7 +15,7 @@ export namespace APIs {
   export import ICorrelationManagementApi = correlationApi.ICorrelationManagementApi;
   export import IEmptyActivityManagementApi = emptyActivityApi.IEmptyActivityManagementApi;
   export import IEventManagementApi = eventApi.IEventManagementApi;
-  export import IFlowNodeManagementApi = flowNodeApi.IFlowNodeManagementApi;
+  export import IFlowNodeInstanceManagementApi = flowNodeInstanceApi.IFlowNodeInstanceManagementApi;
   export import IKpiManagementApi = kpiApi.IKpiManagementApi;
   export import ILoggingManagementApi = loggingApi.ILoggingManagementApi;
   export import IManualTaskManagementApi = manualTaskApi.IManualTaskManagementApi;
